backend/pyproject.toml
[project]
name = "burger_quiz_backend"
version = "0.1.0"
description = "Backend for a BurgerQuiz app"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "gunicorn>=25.0.3",
    "whitenoise>=6.6",
    "psycopg2-binary>=2.9.11",
    "pillow>=12.1.0",
    "djoser>=2.3.3",
    "djangorestframework-simplejwt>=5.3.0",
    "drf-spectacular>=0.29.0",
    "drf-spectacular-sidecar>=2026.1.1",
    "django-taggit>=6.1.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
]
